requirements.txt
pyserial
pillow
pure-python-adb
openai
keyboard
